+++ modules/m_topic.c
@@ -173,7 +173,7 @@ m_topic(struct MsgBuf *msgbuf_p, struct Client *client_p, struct Client *source_
 			sendto_one(source_p, form_str(RPL_TOPICWHOTIME),
 					me.name, source_p->name, chptr->chname,
 					chptr->topic_info,
-					(unsigned long)chptr->topic_time);
+					(long long)chptr->topic_time);
 		}
 	}
 }
